EZP-23978: Split Index RequestEventListener service.
