+++ eZ/Bundle/EzPublishCoreBundle/Tests/EventListener/RequestEventListenerTest.php
@@ -86,47 +86,10 @@ class RequestEventListenerTest extends \PHPUnit_Framework_TestCase
                 KernelEvents::REQUEST => array(
                     array( 'onKernelRequestForward', 10 ),
                     array( 'onKernelRequestRedirect', 0 ),
-                    array( 'onKernelRequestIndex', 40 ),
                 )
             ),
             $this->requestEventListener->getSubscribedEvents()
         );
-    }
-
-    /**
-     * @dataProvider indexPageProvider
-     */
-    public function testOnKernelRequestIndexOnIndexPage( $requestPath, $configuredIndexPath, $expectedIndexPath )
-    {
-        $this->configResolver
-            ->expects( $this->once() )
-            ->method( 'getParameter' )
-            ->with( 'index_page' )
-            ->will( $this->returnValue( $configuredIndexPath ) );
-        $this->request->attributes->set( 'semanticPathinfo', $requestPath );
-        $this->requestEventListener->onKernelRequestIndex( $this->event );
-        $this->assertEquals( $expectedIndexPath, $this->request->attributes->get( 'semanticPathinfo' ) );
-        $this->assertTrue( $this->request->attributes->get( 'needsForward' ) );
-    }
-
-    public function indexPageProvider()
-    {
-        return array(
-            array( '/', '/foo', '/foo' ),
-            array( '/', '/foo/', '/foo/' ),
-            array( '/', '/foo/bar', '/foo/bar' ),
-            array( '/', 'foo/bar', '/foo/bar' ),
-            array( '', 'foo/bar', '/foo/bar' ),
-            array( '', '/foo/bar', '/foo/bar' ),
-            array( '', '/foo/bar/', '/foo/bar/' ),
-        );
-    }
-
-    public function testOnKernelRequestIndexNotOnIndexPage()
-    {
-        $this->request->attributes->set( 'semanticPathinfo', '/anyContent' );
-        $this->requestEventListener->onKernelRequestIndex( $this->event );
-        $this->assertFalse( $this->request->attributes->has( 'needsForward' ) );
     }
 
     public function testOnKernelRequestForwardSubRequest()
